requirements.txt
annotated-types==0.6.0
anyio==4.2.0
ariadne==0.22
blinker==1.7.0
certifi==2023.11.17
click==8.1.7
distro==1.9.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
openai==1.10.0
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic_core==2.16.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.1
tqdm==4.66.1
typing_extensions==4.9.0
Werkzeug==3.0.1
